# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fishing App API
  version: 0.1.0
  description: Fishing App backend API (MVP)
servers:
  - url: http://localhost:3001
    description: Local
  - url: https://api.example.com
    description: Production (placeholder)
...
paths:
...
  /api/catches:
    get:
      tags:
        - Catches
      summary: List catches
      operationId: listCatches
      responses:
        '200':
          description: A list of catches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Catch'
    post:
      tags:
        - Catches
      summary: Create a catch
      operationId: createCatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatchCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/catches/{catchId}:
    parameters:
      - in: path
        name: catchId
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      tags:
        - Catches
      summary: Get a catch
      operationId: getCatch
      responses:
        '200':
          description: Catch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Catches
      summary: Update a catch
      operationId: updateCatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatchUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Catches
      summary: Delete a catch
      operationId: deleteCatch
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/spots:
    get:
      tags:
        - Spots
      summary: List fishing spots
      operationId: listSpots
      responses:
        '200':
          description: A list of spots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spot'
    post:
      tags:
        - Spots
      summary: Create a fishing spot
      operationId: createSpot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spot'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/spots/{spotId}:
    parameters:
      - in: path
        name: spotId
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      tags:
        - Spots
      summary: Get a fishing spot
      operationId: getSpot
      responses:
        '200':
          description: Spot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spot'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Spots
      summary: Update a fishing spot
      operationId: updateSpot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpotUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Spots
      summary: Delete a fishing spot
      operationId: deleteSpot
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: bad_request
            message:
              type: string
              example: Invalid input
            details:
              type: object
              additionalProperties: true
...
    Catch:
      type: object
      required:
        - id
        - species
        - date
      properties:
        id:
          type: integer
          minimum: 1
        species:
          type: string
          example: Robalo
        weight:
          type: number
          format: float
          nullable: true
          example: 2.4
        length:
          type: number
          format: float
          nullable: true
          example: 55
        location:
          type: string
          nullable: true
          example: Praia do Forte
        date:
          type: string
          format: date
        time:
          type: string
          example: '06:15'
        weather:
          type: string
          nullable: true
        baitUsed:
          type: string
          nullable: true
    CatchCreate:
      type: object
      required:
        - species
      properties:
        species:
          type: string
        weight:
          type: number
          format: float
          nullable: true
        length:
          type: number
          format: float
          nullable: true
        location:
          type: string
          nullable: true
        weather:
          type: string
          nullable: true
        baitUsed:
          type: string
          nullable: true
    CatchUpdate:
      type: object
      properties:
        species:
          type: string
        weight:
          type: number
          format: float
          nullable: true
        length:
          type: number
          format: float
          nullable: true
        location:
          type: string
          nullable: true
        weather:
          type: string
          nullable: true
        baitUsed:
          type: string
          nullable: true
...
    Spot:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        catches_count:
          type: integer
          minimum: 0
          nullable: true
        rating:
          type: number
          format: float
          nullable: true
    SpotCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        catches_count:
          type: integer
          minimum: 0
          nullable: true
        rating:
          type: number
          format: float
          nullable: true
    SpotUpdate:
      type: object
      properties:
        name:
          type: string
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        catches_count:
          type: integer
          minimum: 0
          nullable: true
        rating:
          type: number
          format: float
          nullable: true
...
security:
  - bearerAuth: []